Allow ToolPowerAdjust to be in float

+++ Source/EliteBionicsFramework/Util/ToolPowerAdjustInfo.cs
@@ -2,7 +2,7 @@
 {
     public struct ToolPowerAdjustInfo()
     {
-        public int linearAdj = 0;
+        public float linearAdj = 0;
         public float scalingAdj = 1;
 
         public readonly bool HasAdjustment => linearAdj != 0 || scalingAdj != 1;
